requirements.txt
openai>=1.0.0
gradio>=4.0.0
python-dotenv>=1.0.0
pypdf>=3.0.0
tiktoken>=0.5.0
faiss-cpu>=1.7.0
numpy>=1.21.0
